refactor: give the shim record and the provisioning step their own homes
Two units had grown past what one reading holds.

`cli_args/shim.rs` was the `pnpm shim` command *and* the `globalShims`
record it edits — 16 functions over 459 lines. The record moves to
`shim/policy.rs`, where the rules that govern writing it can be stated
once at the top: a shim is not recorded when a layer above the record
would leave it switched off, and a removal never switches a shim on. Both
are rules this PR had to learn the hard way, and neither was visible from
the command that depends on them.

`prepare_package` had grown to 160 lines by acquiring the step that puts
a package manager on the build's PATH. That step is now
`provide_wanted_pm`, which owns the whole decision — pin, host, embedded
pnpm, failure — and hands back the scratch directory the caller keeps
alive. The function reads as its steps again.

No behavior change; the suites and the gate cover it.
